aur-chroot: A local repo without a file:// entry is not an error
`aur sync -c -d foo -r foo_db_path` will call `aur chroot -d foo`, and
if the repo 'foo' does not have a corresponding 'file://' entry in
pacman.conf, then the grep -Em1 fails which exit the script.

But this is not an error, it just means that the corresponding directory
won't be bind_mounted. So add a '|| true' guard to deactivate errexit
for this case.