aur.1: remove bind mount example
The main purpose for using bind mounts is read-only mount. Latter will
prevent use of paccache or pacman -Sc entirely to remove stale
packages, unless specifically pointed to the original repository path
with -c <repo>. If the purpose of a read-only mount was to avoid typos
resulting in data loss, the user could equally typo the -c <repo>
command.