aur-repo: add --table
When this option is used, change the format used by --list
from 'pkgname\tpkgver\n' to 'pkgname\tpkgbase\tpkgver\n'

This assumes that fields in 'desc' files are written in the same order
as repo-add writes them. Especifically: %NAME%, %BASE%, %VERSION%.

If for some reason %NAME% or %BASE% is missing, `-` is shown.
If %VERSION% is missing, that entry is not printed.