man: fix example using escapes incorrectly
An example in aur-graph.1 was not being correcly rendered since it
contained the \n sequence (groff's escape for interpolating registers)
when it wanted to show \n literally.

Use \e to represent a backslash output glyph, which is the recomendation
from man-pages(7).

For consistency, change it on other manual pages as well.