aur-build: add --rm-stale-cache
The issue of checksum mismatches between packages in the local
repository, and packages in a CacheDir is as old as aurutils itself.
Assuming the CacheDir and local repository are on the same machine, the
following workarounds are possible:

1. Bump pkgrel on package rebuilds to avoid mismatches entirely. This
   approach can be automated to a certain degree, e.g. using setconf.
2. Define the local repository to be a CacheDir. This avoids a copy of
   the built package on pacman -S, but may cause packages to be removed
   unexpectdly by tools such as paccache.
3. Create symbolic links in CacheDir using the contents from aur-build
   --results (post-build). This approach is comparable to repose --pool
   <CacheDir>, but may result in broken symlinks if packages are removed
   from the repository.

This commit adds a fourth approach, where it is checked if a built
package is already present in a CacheDir (different from the local
repository root). If so, it is removed, and pacman will have to retrieve
a new copy on pacman -S avoiding any mismatches.

Problematic here is that CacheDir directories (in particular the default
/var/cache/pacman/pkg) may not be writable by users running aur-build.

TODO: Both this approach and 3. can be defined in a seperate script or
aur-repo by using aur-build --results.