Allow overriding pacman.conf for repo selection
The ``pacman.conf`` file might specify an alternate location for your
packages repository – for example, it might specify the ``file://`` path
while you usually use a path on the internet. This is for example useful
if you use the same ``pacman.conf`` on a bunch of machines.

This change passes on ``--pacman-conf`` from ``sync`` to the
repo-manipulating commands.