aur-build: extend --user to local builds
This allows in particular to build packages with one user, and GPG sign
them with another. This takes advantage of aur-build signing packages
seperately from makepkg in /var/tmp. gpg(1) and other commands such as
repo-add(1) are unaffected by this flag and run as the local user.

To use --syncdeps with the build user, an appropriate sudoers rule
should be added. This is similar to the existing "sudo -u build"
approach documented in aur-build(1). If desired, --syncdeps can be
implemented seperately by the user, e.g. to allow an "sudo aur-build"
invocation with dependency resolution.