sync-asroot: rewrite in Python for more argument flexibility
Hard-coding single-letter arguments that are individually forwarded to
either the aur-sync or aur-build calls does not scale well and I was
missing a few, notably --pkgver for aur-build.

Instead of adding yet another single-letter argument, I wanted to make
argument parsing more generic and since I'm not that much into bash, I
converted the script into Python and used some special logic to parse
multiple sections of arguments that are then forwarded to either the
aur-sync call, the aur-build call, or both. Since we already have an
example script in Python, I assume this not to be an issue.

Outside of this, functionality remains largely unchanged.

Options `-k` and `-U` are kept since they are required for the wrapper's
functionality.

All other options can be specified in either the aur-sync section
(started by `--S`) or the aur-build section (started by `--B`). Package
names *must* be specified after a `--` separator argument in order to
support forwarding unrecognized arguments to both aur sub-commands,
e.g. `-d repo`.