aur-fetch: support mixed url and pkgbase names
Support both pkgbase names and urls given via stdin or on the
command-line.

This is done by prefixing pkgbases with AUR_LOCATION and capturing the
basename of paths from urls. Each pkgbase and url is then associated
together so that the main loop can use the remote url specified for each
given package.

This commit introduces a requirement on trurl(1) which is used to
accurately parse out url path components.