graph: detect sections on keywords
Priorly, aur-graph would detect if a pkgname section by the presence of
a blank line. However, these are removed by pacini, causing dependencies
from split packages to be incorrectly considered.

Specifically, when a split package contains an empty dependency field
(depends =) an exit condition is triggered.

The correct approach is to delimit by pkgbase or pkgname keywords, as
described in SRCINFO(5).

Fixes #1228